hwBIG: change update and solve potential issue with size_t - size_t

+++ hwBIG/include/AST.hpp
@@ -18,7 +18,7 @@ struct Node
     virtual ~Node() = default;
 
     virtual std::shared_ptr<Node> clone() const = 0;
-    virtual void updateReferences(const std::pair<int, int>& moveDistance) = 0;
+    virtual void updateReferences(const std::pair<long long, long long>& moveDistance) = 0;
 
     virtual CellValue evaluate() const = 0;
 
@@ -35,7 +35,7 @@ struct CellValueNode : public Node
     {}
 
     std::shared_ptr<Node> clone() const override { return std::make_shared<CellValueNode>(*this); }
-    void updateReferences(const std::pair<int, int>& moveDistance) override {}
+    void updateReferences(const std::pair<long long, long long>& moveDistance) override {}
 
     CellValue evaluate() const override { return value; }
 
@@ -59,7 +59,7 @@ struct CellReferenceNode : public Node
     // expects the clone to be used inside its original spreadsheet 
     std::shared_ptr<Node> clone() const override { return std::make_shared<CellReferenceNode>(*this); }
 
-    void updateReferences(const std::pair<int, int>& moveDistance) override;
+    void updateReferences(const std::pair<long long, long long>& moveDistance) override;
 
     CellValue evaluate() const override;
 
@@ -78,7 +78,7 @@ struct BinaryOperatorNode : public Node
           first(first), second(second)
     {}
 
-    void updateReferences(const std::pair<int, int>& moveDistance) override { first->updateReferences(moveDistance); second->updateReferences(moveDistance); }
+    void updateReferences(const std::pair<long long, long long>& moveDistance) override { first->updateReferences(moveDistance); second->updateReferences(moveDistance); }
 
 protected:
     std::shared_ptr<Node> first;
@@ -92,7 +92,7 @@ struct UnaryOperatorNode : public Node
           first(first)
     {}
 
-    void updateReferences(const std::pair<int, int>& moveDistance) override { first->updateReferences(moveDistance); }
+    void updateReferences(const std::pair<long long, long long>& moveDistance) override { first->updateReferences(moveDistance); }
 
 protected:
     std::shared_ptr<Node> first;
